Get FollowSafely controller to compile. Get Road controller to compile with lots of subordinate controllers   stubbed out. FollowLane controller still not done, so Run controller does not   instantiate Road controller yet.

+++ stacks/art_vehicle/art_nav/src/navigator/follow_safely.h
@@ -23,10 +23,10 @@ public:
 
 private:
   // .cfg variables
-  float close_stopping_distance;
-  float desired_following_time;
-  float max_following_time;
-  float min_following_time;
+  double close_stopping_distance;
+  double desired_following_time;
+  double max_following_time;
+  double min_following_time;
 
   // adjust speed to maintain a safe following time
   void adjust_speed(pilot_command_t &pcmd, float obs_dist);
